requirements.txt
antlr4-python3-runtime==4.7.2
python-telegram-bot==13.1
pillow==8.1.0